change entities , reomve rebase, but comment files and a lot module, user endpoints functional and solve but not docs and tasks

+++ src/main/java/com/saludSystem/repositories/RoleRepository.java
@@ -5,11 +5,14 @@ import com.saludSystem.enums.UserRole;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-
 import java.util.Optional;
+import java.util.UUID;
 
 @Repository
-public interface RoleRepository extends JpaRepository<Role, Integer> {
+public interface RoleRepository extends JpaRepository<Role, Long> {
 
     Optional<Role> findByName(UserRole name);
+
+    //Optional<Role> findById_RoleId(UUID roleId);
+
 }
